refactor: put scripts outside of eleventy.js
