+++ .eleventy.js
@@ -1,6 +1,4 @@
 module.exports = (eleventyConfig) => {
-  const site = require("./src/_data/site")();
-
   eleventyConfig.addPassthroughCopy("src/img");
   eleventyConfig.addPassthroughCopy("src/favicon.ico");
   eleventyConfig.addPassthroughCopy("src/.nojekyll");
@@ -10,34 +8,17 @@ module.exports = (eleventyConfig) => {
   eleventyConfig.setBrowserSyncConfig({
     server: "public",
     port: "3000",
-    startPath: "/a11y-guidelines"
+    startPath: "/a11y-guidelines",
   });
 
-  /**
-   * @param {string} guidelineNumber
-   * @return {[]}
-   */
-  const filterByGuidelineNumber = (post, guidelineNumber) => {
-    return post.filter(p => p.data.number.startsWith(guidelineNumber));
-  }
+  const filterByGuidelineNumber = require("./src/_filters/filterByGuidelineNumber");
   eleventyConfig.addFilter("filterByGuidelineNumber", filterByGuidelineNumber);
 
-  /**
-   * @param {string} scNumber
-   * @returns {string}
-   */
-  const scNumberToPath = (scNumber) => {
-    return scNumber.replace(/\./g, "/");
-  }
+  const scNumberToPath = require("./src/_filters/scNumberToPath");
   eleventyConfig.addFilter("scNumberToPath", scNumberToPath);
 
-  const label = (content, label) => {
-    return `<span class="RuleLabel RuleLabel--${label}">
-<img src="${ site.dir }/img/icon/${label}.svg" width="24" height="24" alt="" />
-${label.charAt(0).toUpperCase() + label.slice(1)}
-</span>${content}`
-}
-  eleventyConfig.addPairedShortcode("label", label);
+  const labelShortCode = require("./src/_shortcodes/label");
+  eleventyConfig.addPairedShortcode("label", labelShortCode);
 
   return {
     dir: {
