+++ src/_filters/scNumberToPath.js
@@ -1,0 +1,9 @@
+/**
+ * @param {string} scNumber
+ * @returns {string}
+ */
+const scNumberToPath = (scNumber) => {
+  return scNumber.replace(/\./g, "/");
+};
+
+module.exports = scNumberToPath;
